@@ -1,10 +1,10 @@
 package com.exampleepam.restaurant.controller;
 
-import com.exampleepam.restaurant.dto.DishResponseDto;
-import com.exampleepam.restaurant.dto.OrderCreationDto;
+import com.exampleepam.restaurant.dto.dish.DishResponseDto;
+import com.exampleepam.restaurant.dto.order.OrderCreationDto;
 import com.exampleepam.restaurant.security.AuthenticatedUser;
 import com.exampleepam.restaurant.service.DishService;
-import com.exampleepam.restaurant.service.UserService;
+import com.exampleepam.restaurant.entity.paging.Paged;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
@@ -12,51 +12,54 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import java.util.List;
-import java.util.Objects;
-
 /**
  * Dish Controller for Users
  */
 @Controller
-public class DishController {
-    private final DishService dishService;
+public class DishController extends BaseController {
 
-    private static final String DEFAULT_SORT_FIELD = "name";
-    private static final String DEFAULT_SORT_DIR = "asc";
-    private static final String DEFAULT_CATEGRY = "all";
+  private final DishService dishService;
 
-    @Autowired
-    public DishController(DishService dishService ) {
-        this.dishService = dishService;
-    }
+  private static final String DEFAULT_SORT_FIELD = "name";
+  private static final String DEFAULT_SORT_DIR = ASCENDING_ORDER_SORTING;
+  private static final String DEFAULT_CATEGORY = "all";
+  private static final int DEFAULT_PAGE = 1;
+  private static final int DEFAULT_PAGE_SIZE = 6;
 
-    @GetMapping("/menu")
-    public String returnMenuSorted(
-            @RequestParam(value = "sortField",
-                    required = false, defaultValue = DEFAULT_SORT_FIELD) String sortField,
-            @RequestParam(value = "sortDir",
-                    required = false, defaultValue = DEFAULT_SORT_DIR) String sortDir,
-            @RequestParam(value = "filterCategory",
-                    required = false, defaultValue = DEFAULT_CATEGRY) String filterCategory,
-            Model model) {
+  @Autowired
+  public DishController(DishService dishService) {
+    this.dishService = dishService;
+  }
 
 
-        List<DishResponseDto> dishes;
+  @GetMapping("/menu")
+  public String returnMenuSorted(
+      @AuthenticationPrincipal AuthenticatedUser user,
+      @RequestParam(value = SORT_FIELD_PARAM,
+          required = false, defaultValue = DEFAULT_SORT_FIELD) String sortField,
+      @RequestParam(value = SORT_DIR_PARAM,
+          required = false, defaultValue = DEFAULT_SORT_DIR) String sortDir,
+      @RequestParam(value = FILTER_CATEGORY_PARAM,
+          required = false, defaultValue = DEFAULT_CATEGORY) String filterCategory,
+      @RequestParam(value = PAGE_NUMBER_PARAM,
+          required = false, defaultValue = "" + DEFAULT_PAGE) int pageNo,
+      @RequestParam(value = PAGE_SIZE_PARAM,
+          required = false, defaultValue = "" + DEFAULT_PAGE_SIZE) int pageSize,
+      Model model) {
 
-        if (Objects.equals(filterCategory, DEFAULT_CATEGRY)) {
-            dishes = dishService.findAllDishesSorted(sortField, sortDir);
-        } else {
-            dishes = dishService.findDishesByCategorySorted(sortField, sortDir, filterCategory);
-        }
+    Paged<DishResponseDto> dishPaged = dishService.findPaginated(pageNo, pageSize,
+        sortField, sortDir, filterCategory);
 
-        model.addAttribute("filterCategory", filterCategory);
-        model.addAttribute(new OrderCreationDto());
-        model.addAttribute("dishList", dishes);
-        model.addAttribute("sortField", sortField);
-        model.addAttribute("sortDir", sortDir);
-        model.addAttribute("reverseSortDir", sortDir.equals("asc") ? "desc" : "asc");
-        return "menu";
-    }
-
+    model.addAttribute(FILTER_CATEGORY_PARAM, filterCategory);
+    model.addAttribute(CURRENT_PAGE_PARAM, pageNo);
+    model.addAttribute(PAGE_SIZE_PARAM, pageSize);
+    model.addAttribute(new OrderCreationDto());
+    model.addAttribute(DISH_LIST_ATTRIBUTE, dishPaged);
+    model.addAttribute(SORT_FIELD_PARAM, sortField);
+    model.addAttribute(SORT_DIR_PARAM, sortDir);
+    model.addAttribute(REVERSE_SORT_DIR_PARAM,
+        sortDir.equals(ASCENDING_ORDER_SORTING) ? DESCENDING_ORDER_SORTING
+            : ASCENDING_ORDER_SORTING);
+    return MENU_PAGE;
+  }
 }
