@@ -1,16 +1,26 @@
 package com.exampleepam.restaurant.entity;
 
-import lombok.*;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
 
-import javax.persistence.*;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.FetchType;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
 import java.math.BigDecimal;
-import java.time.Instant;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 /**
  * Describes Order entity
@@ -22,9 +32,11 @@ import java.util.Objects;
 @Setter
 @Table(name = "orders")
 @ToString
-public class Order extends AbstractBaseEntity{
+public class Order extends AbstractBaseEntity {
 
-    @Column(length = 35)
+    // Increase to 50 so it matches validation constraints in OrderCreationDto
+    // and avoids DataIntegrityViolationException when user enters longer addresses
+    @Column(length = 50)
     private String address;
     private LocalDateTime creationDateTime;
     private LocalDateTime updateDateTime;
@@ -37,17 +49,25 @@ public class Order extends AbstractBaseEntity{
     @JoinColumn(name = "user_id")
     private User user;
 
+    @OneToOne
+    @JoinColumn(name = "order_id")
+    private Review review;
+
+    /**
+     * Indicates whether the customer already submitted reviews for this order.
+     * Used to prevent multiple review submissions.
+     */
+    @Column(nullable = false)
+    private boolean reviewed = false;
+
     @OneToMany(
             mappedBy = "order",
+            fetch = FetchType.EAGER,
             cascade = CascadeType.ALL,
             orphanRemoval = true
     )
     @ToString.Exclude
     private List<OrderItem> orderItems;
-
-
-
-
 
     public Order(long id, Status status, BigDecimal totalPrice,
                  String address, User user) {
@@ -59,6 +79,21 @@ public class Order extends AbstractBaseEntity{
         this.address = address;
         this.user = user;
         this.orderItems = new ArrayList<>();
+    }
+
+    public Order(Long id, String address, LocalDateTime creationDateTime, LocalDateTime updateDateTime,
+                 BigDecimal totalPrice, Status status, User user, Review review,
+                 List<OrderItem> orderItems, boolean reviewed) {
+        this.id = id;
+        this.address = address;
+        this.creationDateTime = creationDateTime;
+        this.updateDateTime = updateDateTime;
+        this.totalPrice = totalPrice;
+        this.status = status;
+        this.user = user;
+        this.review = review;
+        this.orderItems = orderItems;
+        this.reviewed = reviewed;
     }
 
     public void addOrderItem(OrderItem orderItem) {
