@@ -1,33 +1,36 @@
 package com.exampleepam.restaurant.entity;
 
-import lombok.*;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
-
-import javax.persistence.*;
-import java.util.Objects;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 /**
  * Describes OrderItem entity
  */
-@Entity
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class OrderItem extends AbstractBaseEntity{
+@Entity
+public class OrderItem extends AbstractBaseEntity {
 
-    @Column(length = 30)
-    private String dishName;
-    private Integer quantity;
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "dish_id")
+  private Dish dish;
+  private Integer quantity;
 
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "order_id")
-    private Order order;
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "order_id")
+  private Order order;
 
-    public OrderItem(String dishName, Integer quantity) {
-        this.dishName = dishName;
-        this.quantity = quantity;
-    }
+  public OrderItem(Dish dish, Integer quantity) {
+    this.dish = dish;
+    this.quantity = quantity;
+  }
 }
