@@ -22,6 +22,7 @@ public class Paging {
     private boolean prevEnabled;
     private int pageSize;
     private int pageNumber;
+    private int totalPages;
 
     private List<PageItem> items = new ArrayList<>();
 
@@ -35,15 +36,14 @@ public class Paging {
         }
     }
 
-    public void last(int pageSize) {
+    public void last(int totalPages) {
         items.add(PageItem.builder()
                 .active(false)
                 .pageItemType(PageItemType.DOTS)
                 .build());
-
         items.add(PageItem.builder()
                 .active(true)
-                .index(pageSize)
+                .index(totalPages)
                 .pageItemType(PageItemType.PAGE)
                 .build());
     }
@@ -54,7 +54,6 @@ public class Paging {
                 .index(1)
                 .pageItemType(PageItemType.PAGE)
                 .build());
-
         items.add(PageItem.builder()
                 .active(false)
                 .pageItemType(PageItemType.DOTS)
@@ -63,6 +62,7 @@ public class Paging {
 
     public static Paging of(int totalPages, int pageNumber, int pageSize) {
         Paging paging = new Paging();
+        paging.setTotalPages(totalPages);  // totalPages is now set here
         paging.setPageSize(pageSize);
         paging.setNextEnabled(pageNumber != totalPages);
         paging.setPrevEnabled(pageNumber != 1);
@@ -70,21 +70,17 @@ public class Paging {
 
         if (totalPages < PAGINATION_STEP * 2 + 6) {
             paging.addPageItems(1, totalPages + 1, pageNumber);
-
         } else if (pageNumber < PAGINATION_STEP * 2 + 1) {
             paging.addPageItems(1, PAGINATION_STEP * 2 + 4, pageNumber);
             paging.last(totalPages);
-
         } else if (pageNumber > totalPages - PAGINATION_STEP * 2) {
             paging.first(pageNumber);
             paging.addPageItems(totalPages - PAGINATION_STEP * 2 - 2, totalPages + 1, pageNumber);
-
         } else {
             paging.first(pageNumber);
             paging.addPageItems(pageNumber - PAGINATION_STEP, pageNumber + PAGINATION_STEP + 1, pageNumber);
             paging.last(totalPages);
         }
-
         return paging;
     }
 }
