@@ -5,11 +5,9 @@ import com.exampleepam.restaurant.entity.Status;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -28,7 +26,7 @@ public interface OrderRepository extends JpaRepository<Order, Long> {
             @Param(value = "id") long id,
             @Param(value = "statusList") List<Status> statusList, Pageable pageable);
 
-    @Query("SELECT o FROM Order o WHERE o.status = :status AND o.id = :id")
-    Page<Order> findOrdersByStatusAndUserId(@Param(value = "status") Status status,
+//    @Query("SELECT o FROM Order o WHERE o.status = :status AND o.id = :id")
+Page<Order> findOrdersByStatusAndUserId(@Param(value = "status") Status status,
                                             @Param(value = "id") long id, Pageable pageable);
 }
