@@ -1,6 +1,6 @@
 package com.exampleepam.restaurant.validator;
 
-import com.exampleepam.restaurant.dto.UserCreationDto;
+import com.exampleepam.restaurant.dto.user.UserCreationDto;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
